+++ src/components/place/PlaceSelectedListItem.tsx
